+++ lib/product_routes.rb
